refactor: split codegen call helpers
